+++ R/oncoplot.R
@@ -19,7 +19,7 @@
 #' @importFrom aplot insert_top insert_right insert_bottom
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
 #' laml.clin <- system.file('extdata', 'tcga_laml_annot.tsv', package = 'maftools')
 #' laml <- maftools::read.maf(maf = laml.maf, clinicalData = laml.clin)